% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcart.R
\name{bcart}
\alias{bcart}
\title{B-CART model.}
\usage{
bcart(formula, data, iter = 5000, group_variable, pars, scale = TRUE)
}
\arguments{
\item{formula}{The model formula.}

\item{data}{The data to be used in the modeling.}

\item{iter}{The number of iterations for the MCMC.}

\item{group_variable}{The grouping variable.}

\item{pars}{The hyperparameter set}

\item{scale}{Logical to decide whether to scale y or not}
}
\value{
A list containing:
 sigma^2_y, the current errors, the final tree, the
ratios used in the MCMC step and the uniform values sampled,
the final vector of mus and the model formula.
}
\description{
This function runs a BCART model and returns the tree and
other results obtained in the last iteration of the MCMC.
}
\details{
Priors used ----------------------------------------------------------
sigma^2 ~ InvGamma(nu/2, lambda*nu/2), with
the parameters chosen in a way that there is a high probability mass
'before' the RMSE of a linear model

mu_i ~ Normal(mu_mu, sd_mu^2), with
mu_mu = 0
sd_<- = (max(y))/2

Posteriors (both conjugate) ------------------------------------------
sigma^2 ~ InvGamma((nu + n) / 2, 1/2 *(sum_i error^2_i + lambda * nu)))
Note: the error is calculated with the current mu_i for each node

mu_i ~ Normal(mu_post, sigma_mu_post), with
sigma_mu_post = 1/(1/sd_mu + n_in_the_node/sigma^2)
mu_post = (0 + (n_in_the_node/sigma^2) * node.avgResponse* sigma_mu_post
n_in_the_node = observations in each node
node.avgResponse = average of the response in each node
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
