% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman_data.R
\docType{data}
\name{friedman_data}
\alias{friedman_data}
\title{friedman_data
The Friedman data file.}
\format{
A data frame with 5 variables:
\describe{
  \item{\code{y}}{numeric. The response variable.}
  \item{\code{X1}}{numeric. A predictor variable.}
  \item{\code{X2}}{numeric. A predictor variable.}
  \item{\code{X3}}{numeric. A predictor variable.}
  \item{\code{X4}}{numeric. A predictor variable.}
  \item{\code{X5}}{numeric. A predictor variable.}
  }
}
\usage{
data(friedman_data)
}
\description{
friedman_data
The Friedman data file.
}
\keyword{datasets}
