% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybcart-package.R
\docType{package}
\name{mybcart-package}
\alias{mybcart}
\alias{mybcart-package}
\title{mybcart: My mixed-bcart code}
\description{
This package is a very slow implementation of the b-cart model, for didatic purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brunaw}
  \item Report bugs at \url{https://github.com/brunaw}
}

}
\keyword{internal}
