% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_bcart}
\alias{predict_bcart}
\title{Predictions for the B-CART model.}
\usage{
predict_bcart(model, newdata)
}
\arguments{
\item{model}{The model object.}

\item{newdata}{The new data to predict.}
}
\value{
A dataframe with the "prediction" column.
}
\description{
This function predicts for the final tree of a 
Bayesian CART model.
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
