% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{ratio_grow}
\alias{ratio_grow}
\title{Final ratio for a growth step.}
\usage{
ratio_grow(tree, current_node, pars, p, current_selec_var, results_f)
}
\arguments{
\item{tree}{The current tree.}

\item{current_node}{The current grown node.}

\item{pars}{The full list of parameters.}

\item{p}{The number of available predictors}

\item{current_selec_var}{The variable selected for the split.}

\item{results_f}{The current results to find the number of second
generation internal nodes.}
}
\value{
The final ratio for the candidate tree.
}
\description{
The final ratio is to be used as the acceptance
criteria in the MCMC of the b-cart model.
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
