% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ratio.R
\name{transition_ratio_prune}
\alias{transition_ratio_prune}
\title{Transition ratio for prune.}
\usage{
transition_ratio_prune(
  old_tree,
  tree,
  current_node,
  p,
  var_in_prune,
  results_f
)
}
\arguments{
\item{old_tree}{The previous tree.}

\item{tree}{The current tree.}

\item{current_node}{The current node.}

\item{p}{The number of available predictors.}

\item{var_in_prune}{The variable that was split in the node chosen to
be pruned.}

\item{results_f}{The current results to find the number of second
generation internal nodes.}
}
\value{
The transition ratio
}
\description{
Transition probability of going to the candidate tree,
given that the action step was a prune.
}
\details{
When transitioning to a prune, we need the probabilities of:
1. Pruning the tree
2. Selecting  node to prune
When transitioning from a prune back to the split,
we need the probabilities of:
 1. Growing the tree
 2. Growing from the specific pruned node, that has to consider the
 available predictors and available values for that grow.
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
